% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corridor.R
\name{dsr_dalles_requises}
\alias{dsr_dalles_requises}
\title{Selectionner les dalles intersectant une emprise}
\usage{
dsr_dalles_requises(catalogue, emprise)
}
\arguments{
\item{catalogue}{Sortie de \code{\link[=dsr_catalog]{dsr_catalog()}}.}

\item{emprise}{Objet \code{sf} ou \code{sfc}.}
}
\value{
Le catalogue restreint aux dalles concernees.
}
\description{
Selectionner les dalles intersectant une emprise
}
