% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{predict.dsr_modele_conductivite}
\alias{predict.dsr_modele_conductivite}
\alias{print.dsr_modele_conductivite}
\title{Predire une conductivite apprise}
\usage{
\method{predict}{dsr_modele_conductivite}(object, newdata, ...)

\method{print}{dsr_modele_conductivite}(x, ...)
}
\arguments{
\item{object}{Un objet \code{dsr_modele_conductivite}.}

\item{newdata}{\code{SpatRaster} portant au moins les canaux du modele, ou
\code{data.frame} de memes colonnes.}

\item{...}{Ignore.}

\item{x}{Un objet \code{dsr_modele_conductivite}.}
}
\value{
Un \code{SpatRaster} mono-couche \code{p_route}, ou un vecteur numerique selon
la classe de \code{newdata}.
}
\description{
Applique un modele ajuste par \code{\link[=dsr_apprendre_conductivite]{dsr_apprendre_conductivite()}} a une pile de
canaux ou a une table. Sur un \code{SpatRaster}, la sortie est un raster de
probabilite alignee sur l'entree ; les cellules dont un canal manque restent
\code{NA}.
}
\seealso{
\code{\link[=dsr_apprendre_conductivite]{dsr_apprendre_conductivite()}}, \code{\link[=dsr_conductivite]{dsr_conductivite()}}.
}
