% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dessertR-package.R
\docType{package}
\name{dessertR-package}
\alias{dessertR}
\alias{dessertR-package}
\title{dessertR: Correction, Mesure et Qualification de la Desserte Forestiere par Lidar HD}
\description{
Outils pour corriger, mesurer et qualifier un reseau de desserte forestiere a partir des donnees du programme Lidar HD de l'IGN (nuage de points classe, MNT et MNH) et d'une carte de reference vectorielle telle que la BD TOPO. Le package calcule des couches de conductivite separant le signal geomorphologique du signal de surface, repositionne les troncons par recherche de chemin de moindre cout, mesure la geometrie de la plateforme et evalue la praticabilite. Methode adaptee de Roussel et al. (2022) \doi{10.1016/j.jag.2022.103020} au contexte francais.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pobsteta/dessertR}
  \item Report bugs at \url{https://github.com/pobsteta/dessertR/issues}
}

}
\author{
\strong{Maintainer}: Prenom Nom \email{TODO@example.org}

Authors:
\itemize{
  \item Prenom Nom \email{TODO@example.org}
}

}
\keyword{internal}
